backend/requirements.txt
fastapi
fastapi[standard]
uvicorn

SQLAlchemy
psycopg2

python-dotenv
alembic

pydantic[email] 

#hashing            
passlib[bcrypt] 